Clean up old leftovers
